seshat: Remove the types.rs file.
