+++ src/lib.rs
@@ -24,15 +24,11 @@ mod config;
 mod database;
 mod index;
 mod japanese_tokenizer;
-mod types;
 mod events;
 mod error;
 
-pub use database::Connection;
-pub use database::Database;
-pub use database::Searcher;
+pub use database::{Connection, Database, Searcher, SearchResult};
 
-pub use types::SearchResult;
 pub use error::{Result, Error};
 
 pub use config::{Config, Language, SearchConfig};
