--- src/types.rs
@@ -1,43 +1,0 @@
-// Copyright 2019 The Matrix.org Foundation C.I.C.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use std::collections::HashMap;
-use std::sync::mpsc::Sender;
-
-
-use crate::error::Result;
-use crate::events::{Event, HistoricEventsT, Profile, SerializedEvent};
-
-
-pub(crate) enum ThreadMessage {
-    Event((Event, Profile)),
-    HistoricEvents(HistoricEventsT),
-    Write(Sender<Result<()>>),
-}
-
-
-#[derive(Debug, PartialEq, Default, Clone)]
-/// A search result
-pub struct SearchResult {
-    /// The score that the full text search assigned to this event.
-    pub score: f32,
-    /// The serialized source of the event that matched a search.
-    pub event_source: SerializedEvent,
-    /// Events that happened before our matched event.
-    pub events_before: Vec<String>,
-    /// Events that happened after our matched event.
-    pub events_after: Vec<String>,
-    /// The profile of the sender of the matched event.
-    pub profile_info: HashMap<String, Profile>,
-}
